Refactor fingerprinting & packet sniffing
