+++ pyproject.toml
@@ -12,6 +12,7 @@ click = "^8.1.3"
 pyja3 = "^1.0.0"
 psutil = "^5.9.2"
 requests = "^2.28.1"
+mitmproxy = "^8.1.1"
 
 [tool.poetry.scripts]
 benchmark = "proxy_benchmarks.cli:main"
